% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brt_imp.R
\name{Brt_imp}
\alias{Brt_imp}
\title{variable influence of BRT (gbm)}
\usage{
Brt_imp(variabledf, opti = F, ntree = 500, y_varname = c("day_value",
  "night_value", "value_mean"), training, test,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast", ...)
}
\description{
variable influence of BRT (gbm)
}
