% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brt_LUR.R
\name{Brt_LUR}
\alias{Brt_LUR}
\title{Boost regression trees for LUR (using b)}
\usage{
Brt_LUR(variabledf, vis1 = T, y_varname = c("day_value", "night_value",
  "value_mean"), training, test,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast", ...)
}
\arguments{
\item{variabledf}{the dataframe containing predictors and dependent variable}

\item{y_varname}{name of the dependent variable.}

\item{training}{the index for the rows used for training.}

\item{test}{the index for the rows used for testing.}

\item{grepstring}{the variable/column names of predictors in Lasso, grepl stlye, e.g. "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast"}
}
\value{
plot variable importance and an error matrix
}
\description{
Boost regression trees for LUR (using b)
}
