% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree_LUR.R
\name{ctree_LUR}
\alias{ctree_LUR}
\title{Regression tree for LUR (using ctree), mainly for visualisation purpose}
\usage{
ctree_LUR(variabledf, vis1 = T, y_varname = c("day_value",
  "night_value", "value_mean"), training, test,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast", ...)
}
\arguments{
\item{variabledf}{the dataframe containing predictors and dependent variable}

\item{y_varname}{name of the dependent variable.}

\item{training}{the index for the rows used for training.}

\item{grepstring}{the variable/column names of predictors in Lasso, grepl stlye, e.g. "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast"}
}
\value{
plot the tree and return the ctree object
}
\description{
Regression tree for LUR (using ctree), mainly for visualisation purpose
}
