% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_matrix.R
\name{error_matrix}
\alias{error_matrix}
\title{error matrix outputs RMSE, MAE, IQR}
\usage{
error_matrix(validation, prediction)
}
\arguments{
\item{prediction}{The prediction vector}

\item{validataion}{The validataion vector}
}
\value{
an errormatrix with RMSE, MAE, IQR
}
\description{
error matrix outputs RMSE, MAE, IQR
}
