% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_roads.R
\name{merge_roads}
\alias{merge_roads}
\title{create a ring function}
\usage{
merge_roads(inde_var, classvec, keep = F)
}
\arguments{
\item{inde_var}{a dataframe with ROAD variables named "ROAD", 5 types of roads, now is not a generic function, need the data to be formated in a way}

\item{classvec}{the vectors of road types to merge, e.g. c(1,2,3) merges road type 1,2,3.}

\item{keep}{indicate if the road types to be merged are to be kept, default is False, meaning not to keep them}
}
\value{
a data frame with merged road types
}
\description{
create a ring function
}
