% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error.R
\name{plot_error}
\alias{plot_error}
\title{Plot the prediction errors in space}
\usage{
plot_error(validation = 0, prediction = 0, geome, err = NA,
  extremequantil_low = 0.025, extremequantil_high = 0.975,
  plotraster = T)
}
\arguments{
\item{prediction}{The prediction vector}

\item{geome}{geometry of an sf}

\item{err}{The error vector, if provided, the "validataion" and "prediction" are not needed.}

\item{extremequantil_low}{the low quantile to filter out for better visualization}

\item{extremequantil_high}{the high quantile to filter out for better visualization}

\item{validataion}{The validataion vector}
}
\value{
a plot of prediction error distributed in space
}
\description{
Plot the prediction errors in space
}
