% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rsq.R
\name{plot_rsq}
\alias{plot_rsq}
\title{Plot R sqaure with each predictor.}
\usage{
plot_rsq(predictor_matrix, response)
}
\arguments{
\item{predictor_matrix}{a dataframe of predictors}

\item{response}{the response (e.g. air pollution measurement)}
}
\value{
a plot plot individual R square if fitting a linear regression
}
\description{
Plot R sqaure with each predictor.
}
