% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm2ug.R
\name{ppm2ug}
\alias{ppm2ug}
\title{convert ppm to ug/m3}
\usage{
ppm2ug(ppm)
}
\arguments{
\item{df}{values in ppm}
}
\value{
air pollutant in ug/m3
}
\description{
convert ppm to ug/m3
}
\details{
ug/m3 = ppm * 1000 * 1.91
}
