% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDring_coef.R
\name{RDring_coef}
\alias{RDring_coef}
\title{Plot coefficients of roadrings of fitting a linear regression model.}
\usage{
RDring_coef(variabledf, y_var, Road_varname = "ROAD",
  number_roadtypes = 3, buffers_in, buffers_out, pop_var = NA,
  normalize = F)
}
\arguments{
\item{variabledf}{dataframe including predictors and dependent variable}

\item{y_var}{the quote of dependent variable name (e.g. quote (value_mean))}

\item{Road_varname}{the name of variables contains road buffers, e.g. "ROAD"}

\item{number_roadtypes}{number of roadtypes, by default 3}

\item{buffers_in}{buffers of the inner ring}

\item{buffers_out}{buffers of the outer ring}

\item{pop_var}{other variables added to linear regression, in a list or vector if multiple. e.g. popvar = c("pop3k", "RSp", "elevation")}
}
\description{
Plot coefficients of roadrings of fitting a linear regression model.
}
