% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledf.R
\name{sampledf}
\alias{sampledf}
\title{devide the data into test and training set, and dataframe for modeling}
\usage{
sampledf(originaldata, fraction = 0.8, country2digit = NA)
}
\arguments{
\item{originaldata}{original dataframe}

\item{fraction, }{fraction for the training}

\item{country2digit}{country code, 2 digit, if NA or not in the database, the world is returned, the sampling is equal fraction per country}
}
\value{
inde_var the matrix containing response and predictors.  NA values are removed, and columes with all values 0 or less than 0 are removed
}
\description{
devide the data into test and training set, and dataframe for modeling
}
\examples{
a = sampledf(merged, fraction = 0.8, "ES")
test = a$test
training = a$training
inde_var = a$indevar
}
