% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Plot scatterplot between the dependent variables and all the predictors}
\usage{
scatterplot(variabledf, y_name, fittingmethod = "lm")
}
\arguments{
\item{variabledf}{the dataframe containing dependent and independent variables}

\item{y_name}{charater, the name of the dependent variable}

\item{fittingmethod}{the method to fit a line, choosing from the "stat_smooth" in gglot2, can also be  lm, glm, gam, loess.}
}
\description{
Plot scatterplot between the dependent variables and all the predictors
}
