% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_grep.R
\name{subset_grep}
\alias{subset_grep}
\title{subset by grep}
\usage{
subset_grep(df, grepstring)
}
\arguments{
\item{df}{the dataframe to subset}

\item{grepstring}{the namestring of the colomns to grab, greplstyle}
}
\description{
subset by grep
}
