% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univar_rsq.R
\name{univar_rsq}
\alias{univar_rsq}
\title{return R sqaure with each predictor.}
\usage{
univar_rsq(predictor_matrix, response)
}
\arguments{
\item{predictor_matrix}{a dataframe of predictors}

\item{response}{the response (e.g. air pollution measurement)}
}
\value{
the dataframe of r square with each predictor
}
\description{
return R sqaure with each predictor.
}
