% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_imp.R
\name{xgboost_imp}
\alias{xgboost_imp}
\title{return variable importance xgboost}
\usage{
xgboost_imp(variabledf, max_depth = 4, eta = 1, nthread = 2,
  nrounds = 10, y_varname = c("day_value", "night_value",
  "value_mean"), training, test,
  grepstring = "ROAD|pop|temp|wind|Rsp|OMI|eleva|coast", ...)
}
\description{
return variable importance xgboost
}
