% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ring.R
\name{create_ring}
\alias{create_ring}
\title{create a ring function}
\usage{
create_ring(inde_var, normalize = F)
}
\arguments{
\item{inde_var}{a dataframe with ROAD variables named "ROAD", 5 types of roads, now is not a generic function, need the data to be formated in a way}

\item{normalize}{if T, devide the ring by the area of ring. Default is faulse}
}
\value{
the road buffers becomes road ring buffers
}
\description{
create a ring function
}
