% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mechanical.R
\name{mechanical}
\alias{mechanical}
\title{mechanical model: use an exponential function to describe the coefficients away from roads}
\usage{
mechanical(variabledf, y_var = c("day_value", "night_value",
  "value_mean"), pop_var = "pop3k", distance_centre, roadtypes = 3,
  buffers_in, buffers_out, training, test, nls2start = NA,
  Road_varname = "ROAD", normalize = F)
}
\arguments{
\item{variabledf}{the dataframe containing predictors and dependent variable}

\item{y_var}{name of the dependent variable.}

\item{pop_var}{the name of an additional variable as a linear term,  usually population withn a buffer, a string.
@param Road_varname the name of variables contains road buffers, e.g. 'ROAD'}

\item{distance_centre}{the distance to centre from each buffer.  (b2-b1)/2 + b1}

\item{training}{the index for the rows used for training.}

\item{test}{the index for the rows used for testing.}

\item{nls2start}{the start value for nls2. if providing an nls2start, the nls2 from nls2 package is used. Details see nls2.}

\item{normalize}{if True, the road ring is normalised by the area (default is false)}
}
\value{
An object of nls
}
\description{
mechanical model: use an exponential function to describe the coefficients away from roads
}
\details{
This method used nls for modelling. This function also prints errormatrix, the exponential model; plot coefficient. The modelling and evaluation should be separated at a later stage, now putting together for exploration only.
}
