% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rsq.R
\name{plot_rsq}
\alias{plot_rsq}
\title{Plot R sqaure with each predictor.}
\usage{
plot_rsq(predictor_matrix = NA, response = NA, rsqdf = NA,
  varname = NA)
}
\arguments{
\item{predictor_matrix}{a dataframe of predictors}

\item{response}{the response (e.g. air pollution measurement)}

\item{rsqdf}{alternatively can also provide a dataframe with rsq and names of the predictors.}

\item{varname}{name of the column that contains predictor names}
}
\value{
the dataframe of  r square with each predictor
}
\description{
Plot R sqaure with each predictor.
}
