% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm2ug.R
\name{ppm2ug}
\alias{ppm2ug}
\title{convert ppm to ug/m3}
\usage{
ppm2ug(ppm)
}
\arguments{
\item{df}{values in ppm}
}
\value{
air pollutant in ug/m3
library(countrycode)
load("C://Users//lu000012//Documents/files/GLOBAL_LUR//countrywithppm.R")
load("merged")
merged$value_mean = ifelse(merged$country%in%countrywithppm,ppm2ug(merged$value_mean), merged$value_mean )
}
\description{
convert ppm to ug/m3
}
